[refactor] 상품 정보 fetch 로직 상위 컴포넌트로 옮김 #71
+ 자식 컴포넌트에서 판매자 정보를 prop 으로 받아서 사용하기 위함
